+++ src/fruits/difficulty_object.rs
@@ -6,14 +6,14 @@ pub(crate) struct DifficultyObject<'o> {
     pub(crate) base: &'o CatchObject,
     pub(crate) last: &'o CatchObject,
 
-    pub(crate) delta: f32,
-    pub(crate) start_time: f32,
+    pub(crate) delta: f64,
+    pub(crate) start_time: f64,
 
     pub(crate) normalized_pos: f32,
     pub(crate) last_normalized_pos: f32,
 
-    pub(crate) strain_time: f32,
-    pub(crate) clock_rate: f32,
+    pub(crate) strain_time: f64,
+    pub(crate) clock_rate: f64,
 }
 
 impl<'o> DifficultyObject<'o> {
@@ -22,7 +22,7 @@ impl<'o> DifficultyObject<'o> {
         base: &'o CatchObject,
         last: &'o CatchObject,
         half_catcher_width: f32,
-        clock_rate: f32,
+        clock_rate: f64,
     ) -> Self {
         let delta = (base.time - last.time) / clock_rate;
         let start_time = base.time / clock_rate;
