+++ src/fruits/movement.rs
@@ -5,23 +5,23 @@ use std::cmp::Ordering;
 const ABSOLUTE_PLAYER_POSITIONING_ERROR: f32 = 16.0;
 const NORMALIZED_HITOBJECT_RADIUS: f32 = 41.0;
 const POSITION_EPSILON: f32 = NORMALIZED_HITOBJECT_RADIUS - ABSOLUTE_PLAYER_POSITIONING_ERROR;
-const DIRECTION_CHANGE_BONUS: f32 = 21.0;
-const SKILL_MULTIPLIER: f32 = 900.0;
-const STRAIN_DECAY_BASE: f32 = 0.2;
-const DECAY_WEIGHT: f32 = 0.94;
+const DIRECTION_CHANGE_BONUS: f64 = 21.0;
+const SKILL_MULTIPLIER: f64 = 900.0;
+const STRAIN_DECAY_BASE: f64 = 0.2;
+const DECAY_WEIGHT: f64 = 0.94;
 
 pub(crate) struct Movement {
     pub(crate) half_catcher_width: f32,
 
     last_player_position: Option<f32>,
     last_distance_moved: f32,
-    last_strain_time: f32,
+    last_strain_time: f64,
 
-    current_strain: f32,
-    current_section_peak: f32,
+    current_strain: f64,
+    current_section_peak: f64,
 
-    pub(crate) strain_peaks: Vec<f32>,
-    prev_time: Option<f32>,
+    pub(crate) strain_peaks: Vec<f64>,
+    prev_time: Option<f64>,
 }
 
 impl Movement {
@@ -51,7 +51,7 @@ impl Movement {
     }
 
     #[inline]
-    pub(crate) fn start_new_section_from(&mut self, time: f32) {
+    pub(crate) fn start_new_section_from(&mut self, time: f64) {
         self.current_section_peak = self.peak_strain(time - self.prev_time.unwrap());
     }
 
@@ -62,7 +62,7 @@ impl Movement {
         self.prev_time.replace(current.start_time);
     }
 
-    pub(crate) fn difficulty_value(&mut self) -> f32 {
+    pub(crate) fn difficulty_value(&mut self) -> f64 {
         let mut difficulty = 0.0;
         let mut weight = 1.0;
 
@@ -77,7 +77,7 @@ impl Movement {
         difficulty
     }
 
-    fn strain_value_of(&mut self, current: &DifficultyObject<'_>) -> f32 {
+    fn strain_value_of(&mut self, current: &DifficultyObject<'_>) -> f64 {
         let last_player_pos = self
             .last_player_position
             .unwrap_or(current.last_normalized_pos);
@@ -89,14 +89,15 @@ impl Movement {
         let dist_moved = pos - last_player_pos;
         let weighted_strain_time = current.strain_time + 13.0 + (3.0 / current.clock_rate);
 
-        let mut dist_addition = dist_moved.abs().powf(1.3) / 510.0;
+        let mut dist_addition = (dist_moved.abs().powf(1.3) / 510.0) as f64;
 
         if dist_moved.abs() > 0.1 {
             if self.last_distance_moved.abs() > 0.1
                 && dist_moved.signum() != self.last_distance_moved.signum()
             {
-                let bonus_factor = dist_moved.abs().min(50.0) / 50.0;
-                let anti_flow_factor = (self.last_distance_moved.abs().min(70.0) / 70.0).max(0.38);
+                let bonus_factor = (dist_moved.abs().min(50.0) / 50.0) as f64;
+                let anti_flow_factor =
+                    (self.last_distance_moved.abs().min(70.0) / 70.0).max(0.38) as f64;
 
                 dist_addition += DIRECTION_CHANGE_BONUS / (self.last_strain_time + 16.0).sqrt()
                     * bonus_factor
@@ -104,8 +105,8 @@ impl Movement {
                     * (1.0 - (weighted_strain_time / 1000.0).powi(3)).max(0.0);
             }
 
-            dist_addition += 12.5 * dist_moved.abs().min(NORMALIZED_HITOBJECT_RADIUS * 2.0)
-                / (NORMALIZED_HITOBJECT_RADIUS * 6.0)
+            dist_addition += (12.5 * dist_moved.abs().min(NORMALIZED_HITOBJECT_RADIUS * 2.0)
+                / (NORMALIZED_HITOBJECT_RADIUS * 6.0)) as f64
                 / weighted_strain_time.sqrt();
         }
 
@@ -120,7 +121,7 @@ impl Movement {
 
             dist_addition *= 1.0
                 + edge_dash_bonus
-                    * ((20.0 - current.last.hyper_dist) / 20.0)
+                    * ((20.0 - current.last.hyper_dist) / 20.0) as f64
                     * ((current.strain_time * current.clock_rate).min(265.0) / 265.0).powf(1.5);
         }
 
@@ -132,12 +133,12 @@ impl Movement {
     }
 
     #[inline]
-    fn peak_strain(&self, delta_time: f32) -> f32 {
+    fn peak_strain(&self, delta_time: f64) -> f64 {
         self.current_strain * strain_decay(delta_time)
     }
 }
 
 #[inline]
-fn strain_decay(ms: f32) -> f32 {
+fn strain_decay(ms: f64) -> f64 {
     STRAIN_DECAY_BASE.powf(ms / 1000.0)
 }
